NeuEventListener changes and dead code removal (#88)
* hide Hypixel reforge stats

* unused import?

* better comments

* 2.1.md 🙂

* 2.1.md 🙂

* minor cleanup

* api

* remove collectionLog and morus and cape.png

* remove FancyPortals and panorama dev cmd
