+++ src/main/java/io/github/moulberry/notenoughupdates/mixins/MixinGuiInventory.java
@@ -1,7 +1,7 @@
 package io.github.moulberry.notenoughupdates.mixins;
 
-import io.github.moulberry.notenoughupdates.NEUEventListener;
 import io.github.moulberry.notenoughupdates.NotEnoughUpdates;
+import io.github.moulberry.notenoughupdates.listener.RenderListener;
 import net.minecraft.client.gui.inventory.GuiInventory;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -13,7 +13,7 @@ public class MixinGuiInventory {
 	@Inject(method = "drawGuiContainerForegroundLayer", at = @At("HEAD"), cancellable = true)
 	protected void drawGuiContainerForegroundLayer(int mouseX, int mouseY, CallbackInfo ci) {
 		if (NotEnoughUpdates.INSTANCE.config.inventoryButtons.hideCrafting ||
-			NEUEventListener.disableCraftingText) {
+			RenderListener.disableCraftingText) {
 			ci.cancel();
 		}
 	}
